# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "go-gin-backend",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/card": {
            "get": {
                "description": "Retrieve all available cards",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "card"
                ],
                "summary": "Retrieve Cards",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer eyJhb...",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 100,
                        "description": "maximum records per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "page of records, starts from 0",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Card"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "Create new Card",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "card"
                ],
                "summary": "Create Card",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer eyJhb...",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "card",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Card"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Card"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    }
                }
            }
        },
        "/card/user/{user_email}": {
            "get": {
                "description": "Retrieve all cards registered under the given User's email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "card"
                ],
                "summary": "Retrieve all Cards of given User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer eyJhb...",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "user's email",
                        "name": "user_email",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Card"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    }
                }
            }
        },
        "/card/{card_id}": {
            "get": {
                "description": "Retrieve card based on its card_id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "card"
                ],
                "summary": "Retrieve specific Card",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer eyJhb...",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "card's id",
                        "name": "card_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Card"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    }
                }
            },
            "put": {
                "description": "Update specific Card",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "card"
                ],
                "summary": "Update a Card",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer eyJhb...",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "card's id",
                        "name": "card_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "card",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Card"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Card"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Card": {
            "type": "object",
            "properties": {
                "card_id": {
                    "type": "string",
                    "example": "4111222233334444"
                },
                "card_pan": {
                    "type": "string",
                    "example": "xyz"
                },
                "card_type": {
                    "type": "string",
                    "example": "super_miles_card"
                },
                "user_email": {
                    "type": "string",
                    "example": "sudo@gabriel.dev"
                },
                "value": {
                    "type": "number",
                    "example": 100
                },
                "value_type": {
                    "type": "string",
                    "example": "miles"
                }
            }
        },
        "models.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}